requirements.txt
requests==2.32.3
nba_api==1.5.2
python-dotenv==1.0.1
